Fix bug in Autocomplete.
In the original version, [<Name "source">] for each type of completer.
This resuled in WebSharper generating four distinct target values (i.e.
source, source1, source2, source3), which caused autocomplete to fail for all
but the first completer (simple array). Furthermore, autocomplete() in
JQueryUI.js expects a callback with two arguments, but the translated (F#)
callback expects an array with two elements. This is fixed by bridging the
call to the F# function.